+++ bolt-modules/boltlib/spec/functions/apply_prep_spec.rb
@@ -34,10 +34,10 @@ describe 'apply_prep' do
     let(:targets) { hostnames.map { |h| Bolt::Target.new(h) } }
     let(:unknown_targets) { targets.reject { |target| target.protocol == 'pcp' } }
     let(:fact) { { 'osfamily' => 'none' } }
-    let(:custom_facts_task) { Bolt::Task.new(name: 'custom_facts_task') }
-    let(:version_task) { Bolt::Task.new(name: 'puppet_agent::version') }
-    let(:install_task) { Bolt::Task.new(name: 'puppet_agent::install') }
-    let(:service_task) { Bolt::Task.new(name: 'service') }
+    let(:custom_facts_task) { Bolt::Task.new('custom_facts_task') }
+    let(:version_task) { Bolt::Task.new('puppet_agent::version') }
+    let(:install_task) { Bolt::Task.new('puppet_agent::install') }
+    let(:service_task) { Bolt::Task.new('service') }
 
     before(:each) do
       applicator.stubs(:build_plugin_tarball).returns(:tarball)
@@ -46,15 +46,15 @@ describe 'apply_prep' do
       targets.each { |t| inventory.set_feature(t, 'puppet-agent', false) }
 
       task1 = mock('version_task')
-      task1.stubs(:task_hash).returns(name: 'puppet_agent::version')
+      task1.stubs(:task_hash).returns('name' => 'puppet_agent::version')
       task1.stubs(:runnable_with?).returns(true)
       Puppet::Pal::ScriptCompiler.any_instance.stubs(:task_signature).with('puppet_agent::version').returns(task1)
       task2 = mock('install_task')
-      task2.stubs(:task_hash).returns(name: 'puppet_agent::install')
+      task2.stubs(:task_hash).returns('name' => 'puppet_agent::install')
       task2.stubs(:runnable_with?).returns(true)
       Puppet::Pal::ScriptCompiler.any_instance.stubs(:task_signature).with('puppet_agent::install').returns(task2)
       task3 = mock('service_task')
-      task3.stubs(:task_hash).returns(name: 'service')
+      task3.stubs(:task_hash).returns('name' => 'service')
       task3.stubs(:runnable_with?).returns(true)
       Puppet::Pal::ScriptCompiler.any_instance.stubs(:task_signature).with('service').returns(task3)
     end
@@ -198,7 +198,7 @@ describe 'apply_prep' do
     let(:hostnames) { %w[pcp://foo pcp://bar] }
     let(:targets) { hostnames.map { |h| Bolt::Target.new(h) } }
     let(:fact) { { 'osfamily' => 'none' } }
-    let(:custom_facts_task) { Bolt::Task.new(name: 'custom_facts_task') }
+    let(:custom_facts_task) { Bolt::Task.new('custom_facts_task') }
 
     before(:each) do
       applicator.stubs(:build_plugin_tarball).returns(:tarball)
@@ -221,7 +221,7 @@ describe 'apply_prep' do
     let(:hostnames) { %w[foo bar] }
     let(:targets) { hostnames.map { |h| Bolt::Target.new(h) } }
     let(:fact) { { 'osfamily' => 'none' } }
-    let(:custom_facts_task) { Bolt::Task.new(name: 'custom_facts_task') }
+    let(:custom_facts_task) { Bolt::Task.new('custom_facts_task') }
 
     before(:each) do
       applicator.stubs(:build_plugin_tarball).returns(:tarball)